lots of changes in xml output

+++ xml/src/mauprpxml.sh
@@ -1,5 +1,30 @@
 #! /bin/bash
 
+##
+##  This script will populate the template file 'xml/procsum-mauprp.xml'
+##  The changes made to the template, are:
+##  1. a table will be inserted, with phase preprocessing statistics
+##
+##  This script is not meant to be used in 'standalone' mode. 
+##+ It is ignited by 'ddprocess'
+##
+##  For this script to work, it is expected that the user provides a
+##+ a MAUPRP summary file as command line arguments (this file is created
+##+ by RNX2SNX via MAUPRPXTR, placed under the 'OUT' directory and named
+##+ as MPR${YR2}${DOY}0.SUM
+##
+##  Expectations:
+##  The following variables need to be set (i.e. exported by parent 
+##+ process)
+##   * tmpd
+##   * XML_TEMPLATES
+##  The (xml) template to be used:
+##   * ${XML_TEMPLATES}/procsum-mauprp.xml
+##
+##  FEB-2015
+##
+
+
 ## argv[1] - >mauprp.sum
 if [ "$#" -ne 1 ]
 then
@@ -7,14 +32,24 @@ then
   exit 1
 fi
 
-## write the forst part from the template
+## check the variables; thry should be set
 if [ -z ${XML_TEMPLATES} ] || [ -z ${tmpd} ]
 then
   echo "Either XML_TEMPLATES or tmpd variables not defined"
   echo "Refusing to create mauprp.xml"
   exit 1
 fi
-head -n 6 ${XML_TEMPLATES}/procsum-mauprp.xml > ${tmpd}/xml/mauprp.xml
+
+## check that the xml template exists
+if ! test -f ${XML_TEMPLATES}/procsum-mauprp.xml
+then
+  echo "Failed to locate template ${XML_TEMPLATES}/procsum-mauprp.xml"
+  echo "Refusing to create mauprp.xml"
+  exit 1
+fi
+
+## copy the first part of the template
+head -n 30 ${XML_TEMPLATES}/procsum-mauprp.xml > ${tmpd}/xml/mauprp.xml
 
 ## start the table
 echo "<table><title>Pre-processing Information</title>" \
@@ -76,6 +111,30 @@ do
   echo "<row>" >> ${tmpd}/xml/mauprp.xml
   LN=()
   IFS=' ' read -a LN <<< `sed -n "${l}p" ${1}`
+  ##  decide on the background color
+  RMS="${LN[7]}"
+  if test "${RMS}" -gt 20
+  then
+    BGCOLOR=CC0000
+  else
+    if test "${RMS}" -gt 17
+    then
+      BGCOLOR=CC6600
+    else
+      if test "${RMS}" -gt 15
+      then
+        BGCOLOR=CCCC00
+      else
+        BGCOLOR=CCFFFF
+      fi
+    fi
+  fi
+  if test "${RMS}" -lt 12
+  then
+    BGCOLOR=FFFFFF
+  fi
+  echo "<?dbhtml bgcolor='#${BGCOLOR}' ?><?dbfo bgcolor='#${BGCOLOR}' ?>" \
+    >> ${tmpd}/xml/mauprp.xml
   for i in "${LN[@]}"
   do
     echo "<entry>${i}</entry>" >> ${tmpd}/xml/mauprp.xml
